+++ src/Database/index.js
@@ -34,13 +34,20 @@ export default class Database {
   #actionsEnabled: boolean
 
   constructor({ adapter, modelClasses, actionsEnabled }: DatabaseProps): void {
+    if (process.env.NODE_ENV !== 'production') {
+      invariant(adapter, `Missing adapter parameter for new Database()`)
+      invariant(
+        modelClasses && Array.isArray(modelClasses),
+        `Missing modelClasses parameter for new Database()`,
+      )
+      invariant(
+        actionsEnabled === true || actionsEnabled === false,
+        'You must pass `actionsEnabled:` key to Database constructor. It is highly recommended you pass `actionsEnabled: true` (see documentation for more details), but can pass `actionsEnabled: false` for backwards compatibility.',
+      )
+    }
     this.adapter = adapter
     this.schema = adapter.schema
     this.collections = new CollectionMap(this, modelClasses)
-    invariant(
-      actionsEnabled === true || actionsEnabled === false,
-      'You must pass `actionsEnabled:` key to Database constructor. It is highly recommended you pass `actionsEnabled: true` (see documentation for more details), but can pass `actionsEnabled: false` for backwards compatibility.',
-    )
     this.#actionsEnabled = actionsEnabled
   }