+++ CHANGELOG.md
@@ -3,10 +3,18 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
-- [Typescript] 
+
+### ⚠️ Breaking
+
+- Deprecated `bool` schema column type is removed -- please change to `boolean`
+- `experimentalSetOnlyMarkAsChangedIfDiffers(false)` API is now removed
+
+### Improvements
+
+- [Typescript] Typing improvements
      - Added 3 missing properties `collections`, `database` and `asModel` in Model type definition.
-     - Removed optional flag on `actionsEnabled` in the Database constructor options since its mandatory since 0.13.0. 
-     - fixed several further typing issues in Model, Relation and lazy decorator     
+     - Removed optional flag on `actionsEnabled` in the Database constructor options since its mandatory since 0.13.0.
+     - fixed several further typing issues in Model, Relation and lazy decorator
 
 
 ## 0.14.1 - 2019-08-31